Initial commit of Boulderbot Beginner
Boulderbot Beginner is a nearly-empty bot meant to be fully customized by
a new user. The Hello plugin comes with one Command, one Trigger, and your
mod/user join/part echos. You are free to delete the command and trigger
but I wouldn't recommend deleting the echos (though functionally they do
nothing in their current state). The Plugin generator will be the only
thing that gets updated, as Beginner is meant to be a barebones bot.